# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mentorix Backend API
  version: 1.0.0
  description: >
    REST API for Mentorix: authentication, exercise catalog, training programs
    (draft/publish workflow), and admin operations.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://mentorix-backend.onrender.com
    description: Render dev
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register trainer account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          headers:
            Set-Cookie:
              description: HttpOnly refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already registered
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/login:
    post:
      tags: [auth]
      summary: Login with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Authenticated
          headers:
            Set-Cookie:
              description: HttpOnly refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate access token using refresh cookie
      operationId: refresh
      parameters:
        - $ref: "#/components/parameters/RefreshCookie"
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Missing or invalid refresh token
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke current refresh session
      operationId: logout
      parameters:
        - $ref: "#/components/parameters/RefreshCookie"
      responses:
        "204":
          description: Logged out
  /auth/logout-all:
    post:
      tags: [auth]
      summary: Revoke all refresh sessions for the user
      operationId: logoutAll
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RefreshCookie"
      responses:
        "204":
          description: All sessions revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [auth]
      summary: Current user profile
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: User not found
    patch:
      tags: [auth]
      summary: Update current user profile name
      operationId: updateMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MePatchRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: User not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RefreshCookie:
      name: mentorix_refresh
      in: cookie
      required: false
      schema:
        type: string
      description: HttpOnly refresh token (name configurable via REFRESH_COOKIE_NAME)
...
  responses:
    BadRequest:
      description: Invalid request
    Unauthorized:
      description: Missing or invalid access token
    TooManyRequests:
      description: Rate limit exceeded
  schemas:
...
    AuthCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72
        name:
          type: string
          description: Optional display name; defaults to empty string
    MePatchRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Display name; empty string clears the name
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_at, user_id, email]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_at:
          type: string
          format: date-time
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
    MeResponse:
      type: object
      required: [user_id, email, name, created_at, roles]
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time
        roles:
          type: array
          items:
            type: string
            enum: [admin, trainer, client]
...